house-keeping
- script renamed for better readability
